move shuffling to core
