+++ beacon-chain/core/helpers/shuffle_test.go
@@ -1,4 +1,4 @@
-package utils
+package helpers
 
 import (
 	"fmt"
