+++ beacon-chain/core/helpers/spectest/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["shuffle_test_format.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/utils/spectest",
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/helpers/spectest",
     visibility = ["//beacon-chain:__subpackages__"],
 )
