+++ beacon-chain/core/helpers/spectest/shuffle_test_format.go
