+++ beacon-chain/utils/BUILD.bazel
@@ -5,25 +5,15 @@ go_library(
     srcs = [
         "clock.go",
         "flags.go",
-        "shuffle.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/utils",
     visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//shared/bytesutil:go_default_library",
-        "//shared/hashutil:go_default_library",
-        "//shared/params:go_default_library",
-        "@com_github_urfave_cli//:go_default_library",
-    ],
+    deps = ["@com_github_urfave_cli//:go_default_library"],
 )
 
 go_test(
     name = "go_default_test",
     size = "small",
-    srcs = [
-        "clock_test.go",
-        "shuffle_test.go",
-    ],
+    srcs = ["clock_test.go"],
     embed = [":go_default_library"],
-    deps = ["//shared/params:go_default_library"],
 )
